medium workable state while apply import and export
