+++ elements.js
@@ -1,0 +1,3 @@
+export const $arenas = document.body.querySelector('.arenas'),
+    $formFight = document.querySelector('.control'),
+    $chat = document.querySelector('.chat');
