+++ randomiser.js
@@ -1,0 +1,1 @@
+export const getRandom = (min, max) => Math.floor(Math.random() * (Math.floor(max) - Math.ceil(min) + 1)) + Math.ceil(min);
